requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
urlextract
wordcloud
nltk
emoji
plotly
Pillow
